# perf: faster broadcasted