Before

```
Total time: 4.13324 s
File: /home/rvd/src/roelofvandijk/tinygrad/tinygrad/tensor.py
Function: _broadcasted at line 579

Line #      Hits         Time  Per Hit   % Time  Line Contents
==============================================================
   579                                             @profile
   580                                             def _broadcasted(self, fxn:Type[Function], other:Union[Tensor, float], reverse:bool=False) -> Tensor:
   581     42921     107249.3      2.5      2.6      dtype = self.dtype if self.dtype != dtypes.bool and self.dtype.__class__ is not ImageDType else dtypes.float32
   582     42921       9421.7      0.2      0.2      x: Tensor = self
   583     42921     172293.4      4.0      4.2      y: Tensor = Tensor(cast(float, other), device=self.device, requires_grad=False, dtype=dtype) if other.__class__ is not Tensor else cast(Tensor, other)
   584     42921       6955.3      0.2      0.2      if reverse: x, y = y, x
   585     29265      33359.8      1.1      0.8      if x.shape == y.shape: return fxn.apply(x, y)
   586                                           
   587     29265      44341.7      1.5      1.1      len_x_shape, len_y_shape = len(x.shape), len(y.shape)
   588     29265      21575.0      0.7      0.5      max_shape = max(len_x_shape, len_y_shape)
   589                                           
   590     29265     133588.6      4.6      3.2      if len_x_shape != max_shape: x = x.reshape((1,) * (max_shape - len_x_shape) + x.shape)
   591     29265     868346.3     29.7     21.0      if len_y_shape != max_shape: y = y.reshape((1,) * (max_shape - len_y_shape) + y.shape)
   592                                           
   593     29265     130988.0      4.5      3.2      shape_ret = tuple([max(x, y) for x, y in zip(x.shape, y.shape)])
   594     29265     454406.2     15.5     11.0      if x.shape != shape_ret: x = x.expand(shape_ret)
   595     29265     628693.5     21.5     15.2      if y.shape != shape_ret: y = y.expand(shape_ret)
   596                                           
   597     29265    1522022.9     52.0     36.8      return fxn.apply(x, y)
```



after
```
Total time: 2.81013 s
File: /home/rvd/src/roelofvandijk/tinygrad/tinygrad/tensor.py
Function: _broadcasted at line 579

Line #      Hits         Time  Per Hit   % Time  Line Contents
==============================================================
   579                                             @profile
   580                                             def _broadcasted(self, fxn:Type[Function], y:Union[Tensor, float], reverse:bool=False) -> Tensor:
   581     42921      14944.6      0.3      0.5      x: Tensor = self
   582     35598      26421.1      0.7      0.9      if not isinstance(y, Tensor):
   583      7323     161688.8     22.1      5.8        y = Tensor(y, device=self.device, requires_grad=False, dtype=self.dtype if self.dtype != dtypes.bool and self.dtype.__class__ is not ImageDType else dtypes.float32)
   584     42921       9336.7      0.2      0.3      if reverse: x, y = y, x
   585     29265      42097.9      1.4      1.5      if (xshape:=x.shape) == (yshape:=y.shape): return fxn.apply(x, y)
   586                                           
   587     29265      25275.2      0.9      0.9      shape_delta = len(xshape) - len(yshape)
   588     15418       4384.5      0.3      0.2      if shape_delta > 0: y = y.reshape((1,) * shape_delta + yshape)
   589     15418     138657.9      9.0      4.9      elif shape_delta < 0: x = x.reshape((1,) * -shape_delta + xshape)
   590     21465      26939.4      1.3      1.0      if (xshape:=x.shape) == (yshape:=y.shape): return fxn.apply(x, y)
   591                                             
   592     21465      92166.8      4.3      3.3      shape_ret = tuple([max(x, y) for x, y in zip(xshape, yshape)])
   593     21465     453807.5     21.1     16.1      if xshape != shape_ret: x = x.expand(shape_ret)
   594     21465     644855.2     30.0     22.9      if yshape != shape_ret: y = y.expand(shape_ret)
   595     21465    1169554.3     54.5     41.6      return fxn.apply(x, y)
```