@tinyb0t 2023-08-21
Changes made in `tinygrad/`:
```
------------------------------------------------------------
files                             insertions       deletions
------------------------------------------------------------
tinygrad/tensor.py                        13              15
------------------------------------------------------------
lines added in the tinygrad folder: -2
```